ファイル名の修正、finalやprivateの追加、Delegateパターンの分離など可読性の向上
