+++ iOSEngineerCodeCheck/SearchGitHubListViewController.swift
@@ -1,0 +1,91 @@
+//
+//  ViewController.swift
+//  iOSEngineerCodeCheck
+//
+//  Created by 史 翔新 on 2020/04/20.
+//  Copyright © 2020 YUMEMI Inc. All rights reserved.
+//
+
+import UIKit
+
+final class SearchGitHubListViewController: UITableViewController {
+    
+    @IBOutlet private weak var searchBar: UISearchBar!
+    
+    private(set) var repo: [[String: Any]] = []
+    
+    private var task: URLSessionTask?
+    private var word: String!
+    private var url: String!
+    private(set) var idx: Int!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+        searchBar.text = "GitHubのリポジトリを検索できるよー"
+        searchBar.delegate = self
+    }
+    
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        
+        if segue.identifier == "Detail"{
+            let dtl = segue.destination as! SearchedGitHubViewController
+            dtl.vc1 = self
+        }
+        
+    }
+    
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return repo.count
+    }
+    
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        
+        let cell = UITableViewCell()
+        let rp = repo[indexPath.row]
+        cell.textLabel?.text = rp["full_name"] as? String ?? ""
+        cell.detailTextLabel?.text = rp["language"] as? String ?? ""
+        cell.tag = indexPath.row
+        return cell
+        
+    }
+    
+    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        // 画面遷移時に呼ばれる
+        idx = indexPath.row
+        performSegue(withIdentifier: "Detail", sender: self)
+    }
+}
+
+extension SearchGitHubListViewController: UISearchBarDelegate {
+    func searchBarShouldBeginEditing(_ searchBar: UISearchBar) -> Bool {
+        // ↓こうすれば初期のテキストを消せる
+        searchBar.text = ""
+        return true
+    }
+    
+    func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
+        task?.cancel()
+    }
+    
+    func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
+        
+        word = searchBar.text!
+        
+        if word.count != 0 {
+            url = "https://api.github.com/search/repositories?q=\(word!)"
+            task = URLSession.shared.dataTask(with: URL(string: url)!) { (data, res, err) in
+                if let obj = try! JSONSerialization.jsonObject(with: data!) as? [String: Any] {
+                    if let items = obj["items"] as? [[String: Any]] {
+                        self.repo = items
+                        DispatchQueue.main.async {
+                            self.tableView.reloadData()
+                        }
+                    }
+                }
+            }
+            // これ呼ばなきゃリストが更新されません
+            task?.resume()
+        }
+    }
+}
